median kernel now uses an int array as it should.  Also correctly allocating the memory space itll need for medianFilter.  Need to make the medianFilter use multiple spaces now

+++ library/imageAnalysis.h
@@ -45,7 +45,7 @@ void equalizeImageWithHist(Image *image, Image *d_equalizedImage, int *h_mapping
 
 void linearFilter(Image *image, Image *output, float *kernel, int kWidth, int kHeight);
 
-void medianFilter(Image *image, Image *output, float *kernel, int kWidth, int kHeight);
+void medianFilter(Image *image, Image *output, int *kernel, int kWidth, int kHeight);
 
 void saltAndPepperNoise(Image *image, Image *output, int level);
